Byrjun á teams.

+++ src/main/java/is/hi/hbv501gteam23/Controllers/TeamController.java
@@ -17,36 +17,55 @@ import java.util.List;
 public class TeamController {
     private final TeamService teamService;
 
-    /*
+
     @GetMapping
     public List<Team> getAllTeams(){
-        return teamService.listAll();
+        return teamService.findAll();
     }
 
     @GetMapping("/{id}")
     public Team getTeam(@PathVariable Long id){
-        return teamService.getById(id);
+        return teamService.findById(id);
     }
 
     @PostMapping
     public Team createTeam(Team team){
+        if(teamService.findByName(team.getName())!=null)
+        {
+            return null;
+        }
         return teamService.create(team);
     }
 
     @DeleteMapping("/{id}")
     public void deleteTeam(@PathVariable Long id){
+        Team team = teamService.findById(id);
+        if(team!=null){
+            //Print ekkert team til að eyða
+            return;
+        }
+        //Liði eytt
         teamService.delete(id);
 
     }
 
     @PutMapping("/{team}")
     public Team updateTeam(@PathVariable Team team){
+        Long id = team.getId();
+        if(id!=null){
+            //Print ekkert team til að uppfæra
+            return null;
+        }
         return teamService.update(team);
     }
 
     @GetMapping("/{venue}")
     public List<Team> getByVenue(@PathVariable Long id){
-        return teamService.getByVenue(id);
+        return teamService.findByVenue(id);
     }
-    */
+
+    @GetMapping("/{country}")
+    public Team getByCountry(@PathVariable String country){
+        return teamService.findByCountry(country);
+    }
 }
